--- react/typings/global.d.ts
@@ -1,9 +1,0 @@
-interface SelectedItem {
-  unitMultiplier: number
-  measurementUnit: string
-  sellers: Array<{
-    commertialOffer: {
-      AvailableQuantity: number
-    }
-  }>
-}
