+++ react/typings/vtex.product-context.d.ts
@@ -1,14 +1,26 @@
 declare module 'vtex.product-context/useProduct' {
   const useProduct: () => ProductContext
   export default useProduct
+}
 
-  interface ProductContext {
+declare module 'vtex.product-context' {
+  export interface ProductContext {
     selectedQuantity: number
     selectedItem?: SelectedItem
+  }
+
+  interface SelectedItem {
+    unitMultiplier: number
+    measurementUnit: string
+    sellers: Array<{
+      commertialOffer: {
+        AvailableQuantity: number
+      }
+    }>
   }
 }
 
 declare module 'vtex.product-context/ProductDispatchContext' {
-  type DispatchFunction = (payload: { type: string; args?: any }) => void
+  type DispatchFunction = (payload: { type: string; args?: object }) => void
   export const useProductDispatch: () => DispatchFunction
 }
