+++ react/typings/vtex.product-summary-context.d.ts
@@ -5,6 +5,16 @@ declare module 'vtex.product-summary-context/ProductSummaryContext' {
     selectedItem?: SelectedItem
   }
 
-  type DispatchFunction = (payload: { type: string; args?: any }) => void
+  interface SelectedItem {
+    unitMultiplier: number
+    measurementUnit: string
+    sellers: Array<{
+      commertialOffer: {
+        AvailableQuantity: number
+      }
+    }>
+  }
+
+  type DispatchFunction = (payload: { type: string; args?: object }) => void
   export const useProductSummaryDispatch: () => DispatchFunction
 }
